organized data related folders structure, updated dependencies accordingly
