--- src/dummydata/dumbdatawithNotes.csv
@@ -1,5 +1,0 @@
-Curso,Seccion,Creditos,Reuniones,Profesor,Grades
-MATE4145,70,4,1:00pm-2:50pm LW AE406 | 1:00pm-1:50pm V AE406,Angel Cruz,A
-INSO4101,40,3,10:30am-11:20am LMW S113,Marko Schutz,B
-INGE3011,96,2,3:30pm-5:20pm MJ S316,Cassandra Lozada Figueroa,C
-INEL4115,66,3,12:30pm-2:20pm M S104A,Kevin Astudillo Senmache,A
