+++ src/pages/studentPage.js
@@ -2,9 +2,9 @@ import React, { useState, useEffect } from "react";
 import styles from "../styles/pages/studentPage.module.scss";
 import Coursetable from "../components/courseTable";
 import Scheduletable from "../components/scheduleTable";
-import Datafall2023 from "../dummydata/fall2023.csv";
-import Dataspring2024 from "../dummydata/spring2024.csv";
-import officeHours from "../dummydata/profOHspring2024.csv";
+import Datafall2023 from "../data/dummy_data/fall2023.csv";
+import Dataspring2024 from "../data/dummy_data/spring2024.csv";
+import officeHours from "../data/dummy_data/profOHspring2024.csv";
 import Examtable from "../components/activeExams";
 import StudentCard from "../components/studentCard";
 import linkBox from "../components/linkBox";
